Telegram bot fix

+++ project/telegram-service/src/main/java/org/example/services/handlers/GuessHandler.java
@@ -7,6 +7,7 @@ import org.example.services.RedisService;
 import org.example.services.UserSessionManager;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
+import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.RestClient;
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.objects.Update;
@@ -39,20 +40,27 @@ public class GuessHandler implements IHandler {
             }
             case WAITING_FOR_GUESS -> {
                 String gameId = redisService.getGameId(update.getMessage().getFrom().getId());
+                try{
+                    if(userMessage != null) {
+                        System.out.println("THis is usermessage:" + userMessage);
+                        ResponseEntity<GameGuessResponseDTO> response = restClient.post()
+                                .uri("/game/" + gameId + "/guess")
+                                .body(userMessage)
+                                .retrieve()
+                                .toEntity(GameGuessResponseDTO.class);
 
-                ResponseEntity<GameGuessResponseDTO> response = restClient.post()
-                        .uri("/game/" + gameId + "/guess")
-                        .body(userMessage)
-                        .retrieve()
-                        .toEntity(GameGuessResponseDTO.class);
-
-
-                if (response.getBody().status().equals("WON")) {
-                    redisService.removeGameID(update.getMessage().getFrom().getId());
+                        if(response.getBody().status().equals("WON")) {
+                            redisService.removeGameID(update.getMessage().getFrom().getId());
+                        }
+                        String hint = response.getBody().hint();
+                        sendMessage.setText(response.getBody().message() + (hint == null? "" : ". Hint: " + hint));
+                    }
+                } catch (HttpClientErrorException e) {
+                    sendMessage.setText("Error: " + e.getResponseBodyAsString());
                 }
-                String hint = response.getBody().hint();
-                sendMessage.setText(response.getBody().message() + (hint == null? "" : ". Hint: " + hint));
-
+                catch (Exception e) {
+                    sendMessage.setText("Error occurred" + e.getMessage());
+                }
                 userSessionManager.endSession(chatId);
             }
             default -> {
